Consolidate OAuth 2.0 logic
This puts the OAuth 2.0 logic unique to the Google integration
into the oauth2 module, so it's the same for both Google and Gitlab.

This also made refresh tokens optional, since they are in the OAuth
2.0 spec.
